V124: deep debt resolution + tolerance hardening
eps module promoted to pub with SAFE_DIV_STRICT, all bare float literals
replaced with named tol::/eps:: constants, NDJSON injection hardening
(RFC 8259 escaping), validate_all exit-code disambiguation (0/1/2),
capability-based NestGate discovery (no hardcoded addresses), UID
discovery 4-tier chain, 29 validator exit-code integration tests, CI
validate-all job, 5 pre-existing clippy fixes. All docs/specs/handoffs
updated to V124, wateringHole handoff crafted.

Made-with: Cursor

+++ crates/groundspring/src/rare_biosphere.rs
@@ -169,7 +169,7 @@ fn abundance_occupancy_gpu(
         .sample(
             &cumulative,
             Some(&mut seeds),
-            crate::cast::u64_u32_truncate(depth as u64),
+            crate::cast::u64_u32_truncate(depth),
             crate::cast::u64_u32_truncate(n_samples as u64),
             config,
         )
@@ -321,7 +321,7 @@ fn tier_detection_rate_gpu(
         .sample(
             &cumulative,
             Some(&mut seeds),
-            crate::cast::u64_u32_truncate(depth as u64),
+            crate::cast::u64_u32_truncate(depth),
             crate::cast::u64_u32_truncate(n_replicates as u64),
             config,
         )
